Add @BusConnectionFactory to qualify the preferred instance
If you are using AMQP there needs to be a ConnectionFactory (from
Spring Rabbit) in the application context. If there is a single
ConnectionFactory it will be used, or if there is a one qualified as
@BusConnectionFactory it will be preferred over others, otherwise
the @Primary one will be used. If there are multiple unqualified
connection factories there will be an error.

Fixes gh-13