Compatibility for independent bus clients (#76)
independent bus clients may not have all events implemented or the location of the event classes could be in a different package.
This PR fixes issues in such an environment (e.g. Microservices). Therefore the event type in the AckRemoteApplicationEvent will be set via a special setter, so the availability of the type can be checked.
In addition the UnknownRemoteApplicationEvent is fixed to not cause a NullPointerException when raised.

Fixes #74