Instead of ApplicationContext ID use BusProperties to address events
The ApplicationContext is not a very good place to store something
that has to be unique per instance across a scaled up app. This change
extracts the old value out into a new spring.cloud.bus.id.

It also adds a unique identifier to the end of the id. Hopefully
this interops with older versions. If not we can change the default
and/or add a property to switch off this change.