Refactors bus to use spring cloud function. (#236)
Rather than the annotation model. The monolithic `BusAutoConfiguration` was split.

BusBridge is now an interface with a send() method. In the future, there will be an RSocketBusBridge, the default is `StreamBusBridge`.

StreamBusBridge uses StreamBridge to send messages.

RemoteApplicationEventListener listens for remote events sends them to bus if needed.

Adds an amqp integration test.

fixes gh-227